backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic>=2.10
python-dotenv>=1.0
supabase>=2.11
python-jose[cryptography]>=3.3
httpx>=0.28
pytest>=8.3
pytest-asyncio>=0.25
python-multipart>=0.0.20
passlib[bcrypt]>=1.7
